// src/pages/Local/LocalMarket/types/localMarket.types.ts
export interface LocalMarketDetailType {
    content: string;
    content_url: string;
    create_at: string;
    update_at: string;
    keyword: string;
    era: string;
    la: number;
    lo: number;
    lcc_address: string;
    main_thumb_url: string;
    rel_rest_address: string;
    rel_rest_name: string;
    rel_rest_tel: string;
    sub_title: string;
    middle_title: string;
    title: string;
  }